Improve WGS84 bounding box range checking
Add method to check for spanning 180 degree meridian.
Move methods into GeometryUtils for reuse.
Check envelope of uploaded OSM and GTFS data.
Check envelope of WebMercatorGridPointSet when created.
Check envelope of StreetLayer when it's loaded.
Text describing range-checked object is now substituted into messages.
